[warnings] fix integer casting
* use asserting casts to detect overflows or sign conversions
* code cleanup for issues uncovered by casts

+++ libfreerdp/core/client.h
@@ -53,8 +53,8 @@ typedef struct
 typedef struct
 {
 	char name[CHANNEL_NAME_LEN + 1];
-	int OpenHandle;
-	int options;
+	uint32_t OpenHandle;
+	ULONG options;
 	int flags;
 	void* pInterface;
 	rdpChannels* channels;
